requirements.txt
discord.py~=2.0
aiohttp~=3.8
beautifulsoup4~=4.11
soupsieve~=2.3
